client functional, linted, and renamed
