+++ components/log-viewer-webui/client/package.json
@@ -28,6 +28,7 @@
     "@eslint/js": "^9.19.0",
     "@types/react": "^19.0.10",
     "@types/react-dom": "^19.0.4",
+    "@types/socket.io-client": "^3.0.0",
     "@vitejs/plugin-react": "^4.3.4",
     "eslint": "^9.19.0",
     "eslint-config-yscope": "latest",
